+++ src/1-one/match/methods/match/02-from-here.js
@@ -1,8 +1,9 @@
-import { getGroup } from './_lib.js'
 import doAstrix from './steps/astrix.js'
 import doOrBlock from './steps/or-block.js'
 import doAndBlock from './steps/and-block.js'
-import { isEndGreedy, getGreedy } from './logic/greedy.js'
+import doNegative from './steps/negative.js'
+import simpleMatch from './steps/simple-match.js'
+import { isEndGreedy } from './logic/greedy.js'
 import matchTerm from './term/doesMatch.js'
 
 
@@ -43,7 +44,6 @@ const tryHere = function (terms, regs, start_i, phrase_length) {
       if (haveNeeds === false) {
         break //done!
       }
-      // log(`✗ |terms done|`)
       return null // die
     }
     // support 'unspecific greedy' .* properly
@@ -70,95 +70,34 @@ const tryHere = function (terms, regs, start_i, phrase_length) {
       }
       continue
     }
+    // support '.' any, but single
+    if (reg.anything === true) {
+      let alive = simpleMatch(state)
+      if (!alive) {
+        return null
+      }
+      continue
+    }
     // ok, finally test the term/reg
-    let term = state.terms[state.t]
-    let hasMatch = matchTerm(term, reg, state.start_i + state.t, state.phrase_length)
-    if (reg.anything === true || hasMatch === true || isEndGreedy(reg, state)) {
-      let startAt = state.t
-      // if it's a negative optional match... :0
-      if (reg.optional && regs[state.r + 1] && reg.negative) {
-        continue
+    let hasMatch = matchTerm(state.terms[state.t], reg, state.start_i + state.t, state.phrase_length)
+    if (hasMatch === true || isEndGreedy(reg, state)) {
+      let alive = simpleMatch(state)
+      if (!alive) {
+        return null
       }
-      // okay, it was a match, but if it's optional too,
-      // we should check the next reg too, to skip it?
-      if (reg.optional && regs[state.r + 1]) {
-        // does the next reg match it too?
-        let nextRegMatched = matchTerm(term, regs[state.r + 1], state.start_i + state.t, state.phrase_length)
-        if (reg.negative || nextRegMatched) {
-          // but does the next reg match the next term??
-          // only skip if it doesn't
-          let nextTerm = state.terms[state.t + 1]
-          if (!nextTerm || !matchTerm(nextTerm, regs[state.r + 1], state.start_i + state.t, state.phrase_length)) {
-            state.r += 1
-          }
-        }
-      }
-      // log(`✓ |matched '${state.terms[state.t].normal}'|`)
-      //advance to the next term!
-      state.t += 1
-      //check any ending '$' flags
-      if (reg.end === true) {
-        //if this isn't the last term, refuse the match
-        if (state.t !== state.terms.length && reg.greedy !== true) {
-          // log(`✗ |end-flag|`)
-          return null //die
-        }
-      }
-      //try keep it going!
-      if (reg.greedy === true) {
-        state.t = getGreedy(state, regs[state.r + 1])
-        if (state.t === null) {
-          // log(`✗ |too-short|`)
-          return null //greedy was too short
-        }
-        if (reg.min && reg.min > state.t) {
-          // log(`✗ |too-short2|`)
-          return null //greedy was too short
-        }
-        // if this was also an end-anchor match, check to see we really
-        // reached the end
-        if (reg.end === true && state.start_i + state.t !== phrase_length) {
-          // log(`✗ |not-end|`)
-          return null //greedy didn't reach the end
-        }
-      }
-      if (state.hasGroup === true) {
-        // Get or create capture group
-        const g = getGroup(state, startAt)
-        // Update group - add greedy or increment length
-        if (state.t > 1 && reg.greedy) {
-          g.length += state.t - startAt
-        } else {
-          g.length++
-        }
-      }
-      // should we clump-in the 2nd word of a contraction?
-      // let lastTerm = state.terms[state.t - 1]
-      // let thisTerm = state.terms[state.t]
-      // if (lastTerm && thisTerm && lastTerm.implicit && thisTerm.implicit) {
-      //   // only if it wouldn't match, organically
-      //   let nextReg = regs[state.r + 1]
-      //   if (!nextReg || !matchTerm(thisTerm, nextReg, state.start_i + state.t, state.phrase_length)) {
-      //     state.t += 1
-      //   }
-      // }
       continue
     }
 
     // ok, it doesn't match.
     // did it *actually match* a negative?
     if (reg.negative) {
-      let tmpReg = Object.assign({}, reg)
-      tmpReg.negative = false // try removing it
-      let foundNeg = matchTerm(state.terms[state.t], tmpReg, state.start_i + state.t, state.phrase_length)
-      if (foundNeg === true) {
-        // log(`✗ |no neg|`)
-        return null //bye!
+      let alive = doNegative(state)
+      if (!alive) {
+        return null
       }
     }
     //bah, who cares, keep going
     if (reg.optional === true) {
-      // log(`- |optional reg '${reg.word}'|`)
       continue
     }
 
@@ -169,7 +108,6 @@ const tryHere = function (terms, regs, start_i, phrase_length) {
       }
       // does the next one match?
       if (matchTerm(state.terms[state.t + 1], reg, state.start_i + state.t, state.phrase_length)) {
-        // log(`✓ |contraction| '${state.terms[state.t + 1].implicit}'`)
         state.t += 2
         continue
       }
@@ -177,16 +115,16 @@ const tryHere = function (terms, regs, start_i, phrase_length) {
     return null //die
   }
   //return our results, as pointers
-  let pntr = [null, start_i, state.t + start_i] //`${start_i}:${state.t + start_i}`
+  let pntr = [null, start_i, state.t + start_i]
   if (pntr[1] === pntr[2]) {
-    // log(`✗ |found nothing|`)
+    //found 0 terms
     return null
   }
   let groups = {}
   Object.keys(state.groups).forEach(k => {
     let o = state.groups[k]
     let start = start_i + o.start
-    groups[k] = [null, start, start + o.length] //`${start}:${start + o.length}`
+    groups[k] = [null, start, start + o.length]
   })
   return { pointer: pntr, groups: groups }
 }
